Move opcode functions inside Machine class
The way they were headed would require them also taking a Machine& parameter, which makes it similar to a non-static member function. It would be redundant if they continued to be non-member functions. Because of that, they were made private and are now controlled by the new execute() function, which will call the appropriate opcode.