Implement zero register in reg()
The zero register is a special register which is enforced by hardware convention. When this register is read, it always produces the value 0. In effect, writing to this register also does virtually nothing. In this implementation, setting reg0 to 0 before returning a reference to it will ensure that reading will always produce the value 0.