Implement jalr opcode
The jalr opcode is similar to the MIPS jump-and-link-register opcode, but with a couple differences. This one stores pc+1 into a register of your choice(MIPS always chose $31 which, by convention, was the return address register), then jumps to the address stored in a register of your choice. It also has an optional immediate field which, if nonzero, causes a syscall instead with the two given registers.