Create lexer function
The lexer() phase is the first phase of a compiler, where it splits everything up into keywords, tokens, etc. and gives them an identifier or category as well. Usually they're implemented with regex, but that will probably be done later as soon as C++11 actually has optimized, hopefully compile-time-generated regexes or until I make my own.