Update Lexer test to changes from lexer.h
Before, the lexer was just a function that returned a vector of Lexemes. With next() yet to be implemented, the lexer is now a class which handles everything on its own without having to pollute the global namespace with a ton of functions that the main program won't even use.