Implement next() in Lexer class
I've decided that lexical analysis should be handled in the constructor instead of happening in short fragments in the next() function. That way, getting the Lexemes in order would be faster after the lexemes queue has been constructed, instead of making a short pause happen for every call to next().