Add space/tab case
When the lexer hits a space or a tab, it will create a new substring between that space and where the base pointer (iterator) left off. The trick is, it doesn't do any copying or anything like that. Just assigns a null terminator to a character in the copy of the file and calls tokenize() since the file contents will only be used once.