Add newline case
When the Lexer hits a newline, it should first check if the token before it hasn't been tokenized yet and will do that before anything else happens. After that, it emplaces a newline Lexeme to terminate the line. Then similar code in the end of space where it resets the iterator and index.