Add optimization
In my state machine implementation, I have this default state that happens when the machine rejects the input. That was only realized by the end of the loop after it checks if it ended up in an accepting state. Now, at the end of the while loop, it checks if it just got into the rejecting state, and if so it will immediately reject the input and return false.