Add analyze() to Lexer class
This member function will be used in place of the constructor's main code, but instead of doing a copy-paste I'm going to redo it with the state machine logic shown in isImmediate(), isRegister(), etc. with a large amount of states and possibilities. The reason why it's being moved to a member function is because it might be useful to have the ability to recurse over itself like in the case of a .include or other assembler directives that may be implemented in the future.