Add imm check states
The states just added are minus check and digit check. An immediate is loosely defined as a word starting with a digit and the rest containing either digits or letters. That way, a mispelled number will be handled by a different stage and reduces the complexity of the lexer's already complex state machine.