Add note to register state
A register name is defined by the character 'r' followed by a digit between 0 and 7. Because it starts with an r and no other states have a special case such as that, it would be easier on memory to just implement a special case that would cause it to enter that state.