Add register check state
When the lexer finds an 'r', it goes to this state. If the next character is a digit, it goes to the reg accepting state, else it gets resolved as an identifier. If the reg accepting state finds a non-comment/whitespace character immediately after, it also gets resolved as an identifier.