Implement default case
This case is slightly more complex than the other special cases. It first sets lastState to currState because all cases modify currState, then it goes through an if/else chain which goes through the regular state columns. The final else will someday display a lexer error to the user and for now will set the state to the error state.