Add parseState() private function to Lexer class
This function takes in a state and parses it to see what kind of token should be produced. If it's a non-accepting state, it displays an error, increments the error count, and returns a Token::Type of ERROR. Else it goes through the rest of the accepting states to see whcih Type it should return.