Implement parseState()
The parseState() static helper function for the Lexer class basically takes in a state that the lexer was in and returns the appropriate Token Type e.g. if the lexer was in an accepting state for an immediate, it would return a Token Type of Immediate. If in a non-accepting state, it returns Error.