Rewrite newline handler
After some deep thought and head-scratching, I found that the only thing that affects the creation of tokens (in this simple assembly language) is whitespace (newline, space, tab, etc), so I'm starting to rewrite it all to handle space/tab and newline separately. The code segment I added first checks if the current state is a newline state. If so, then it will look back exactly one state to figure out how it should react to this newline. After possibly creating a token, it wraps it all up in a newline token and resets the iterator and state variables so the lexer will be prepared to take in the next line of code.