Condense state-parsing segment
I noticed some repetition with the end of word case, and I realized it could be placed inside the first case that checks for a newline. Now, that first if is compounded to include both space/tab and newlines. The compound check with the current state being not equal to the starting state was omitted because it's already implied by the else keyword.