Implement Terminal class
A Terminal symbol just contains a Token and can act as a node of a parse tree, specifically a leaf. However, its value isn't always known the moment it's created. A production from a NonTerminal would have to go to the next input on the next iteration in order to initialize the Terminal, or go off on another path, which was why setToken was added. The getToken function is so other stages can read the parse tree and do stuff with it.