Add ParseTable and Production class
These two classes are going to be used to the Parser's parse table. A parse table maps NonTerminal symbols with Token Types onto a table containing Productions. A Production will basically be an array of ParseNode::Symbol enums which tells a NonTerminal object how to apply this specific production in order to produce the right ParseNodes.