Add parent member to ParseNode
This member, which propagated into a ton of changes, is used for traversing the parse tree and being able to have an iterator go through every leaf of the tree. To be able to go through every leaf, it has to be able to backtrack, and for that to be possible, every node has to have some knowledge of its parent.