Create base iterator class for ParseTree iterators
I was tired of code duplication with iterator and const_iterator, so I made a template class that provides all the functionalities of both classes and made iterator and const_iterator typedefs of that class. Usually it isn't all that possible to define the member functions of a template class outside the header file, but because I declared a typedef which uses that template class, I was safely able to do so at least on my compiler.