Add Operand class tree
Luckily there are only 3 types of operands that can be used in a single instruction: registers, immediates, and identifiers(placeholders for an immediate). Only registers and immediates are used in code generation. The concept of labels and identifiers are left to the assembler to translate into immediates, usually by building a symbol table.