Change queue to vector
If I were to implement next() in Parser in a similar way to Lexer, it would cause a memory leak because of the heap-allocated Lines not being deleted because just the pointers are being freed. Passing Lines by value would slice the label off. Using a vector and an iterator, the pointers can sit in memory and be accessed until the Parser goes out of scope, in which all the Lines are deleted.