Add operand error handling in Parser class
Most of the changes were directed at Immediate's constructor, since stoi can throw an invalid_argument exception if no conversion can be performed or an out_of_range if the value is too big. Both are simplified into an invalid immediate exception which is correctly displayed by the try-catch block above that change.