Fix premature deletion of Operand objects
In assemble(), the function was creating a temporary Line object which contains a copy of the given Line but with identifiers resolved. The problem was that it was copying pointers from the given Line without copying the object. When the temporary line was destroyed, so were some of the pointers the given Line contained, causing many pointers to become dangling pointers which is VERY bad. To resolve this, I made a non-const getOperands() for Line and made assemble() directly modify the given line's operand list in a non-leaking way.