improve memory system
The memory header/source pair was removed an replaced by a templated using
declaration.

The regs and ram member variable types were modified to use that type.

The mem() function was removed along with getReg() and getMem().

The reg() function no longer does bounds checking, and has an overload for
const which returns by value instead.

When making a syscall, an out_of_range exception will be thrown instead of
calling exit(1).