fix minor style inconsistencies
An include for <stdexcept> was added to show that std::out_of_range didn't
just come from nowhere (which I think was included in <array>).

A barrage of newlines was removed from the Bitwise enum definition, since
context and self-documentation can usually tell you what category the enum
values are being grouped as.

The line of comments that state what the includes are for were reindented
since I forgot to do that.

Default construction of reg and ram members is already implied in the
constructor so the explicit calling of their constructors was omitted for
the sake of simplicity since there was nothing inside those little curly
braces to begin with.

The Memory using declaration was moved to be below the 2 typedefs since
they needed to be defined before Memory could be declared.