improve error handling
The std::range_error exception class was used instead of std::out_of_range
because out of range refers to an argument of a function, whereas a range
error refers to an internal computation of a function rather than its
argument. If execute() were split up into functions, then it would be
appropriate to use the std::out_of_range exception class.

The invalid syscall id exception was moved to the correct spot and an
invalid opcode exception was added, but expected to never happen.