split up execution into multiple functions
For minor self-documentation and style purposes, execute() was renamed to
start().

A private execute() function was added instead, which is meant to execute
a single instruction.

The new execute() was then split up into 8 smaller functions, each
corresponding to a RiSC opcode.

A syscall() function was added for use in the jalr() function as well.