fix bytecode loading
For char's that have the most significant bit set, applying a bitwise
operation extended the sign while it was being promoted to an int. To
prevent the sign extension, a cast to unsigned char was added. If this
change weren't made, an "addi r1 r1 -1", when loaded, would appear to have
the bytecode 0xFFFF.