fix code generation in second pass
The second pass of the assembler is when it tries to go through the lines of
code that previously failed to compile given an incomplete symbol table,
this time with a complete one. There was a very small error in my code,
which used the in-scope line variable (which is left at the end of the
input) as the parameter, rather than the saved line value.